requirements.txt
flask
pillow
pytesseract
textblob
